Revert "move tests and add new for sqlalchemyerror"
This reverts commit 11b4716f29fffdada9ae2bea0618af255f76d4a7.
